move the function into internal not protected
