+++ include/ginkgo/core/config/registry.hpp
@@ -213,23 +213,6 @@ inline const CriterionFactoryMap& registry::get_map_impl<CriterionFactoryMap>()
 }
 
 
-template <typename T>
-struct type_string {
-    static std::string str() { return "N"; };
-};
-
-#define TYPE_STRING_OVERLOAD(_type, _str)         \
-    template <>                                   \
-    struct type_string<_type> {                   \
-        static std::string str() { return _str; } \
-    }
-
-TYPE_STRING_OVERLOAD(double, "double");
-TYPE_STRING_OVERLOAD(float, "float");
-TYPE_STRING_OVERLOAD(std::complex<double>, "complex<double>");
-TYPE_STRING_OVERLOAD(std::complex<float>, "complex<float>");
-
-
 }  // namespace config
 }  // namespace gko
